modularizing the authetication
