+++ Backend/.eslintrc.js
@@ -1,0 +1,9 @@
+module.exports = {
+    // Other ESLint configurations...
+  
+    rules: {
+      'no-unused-vars': ['warn', { 'vars': 'all', 'args': 'none' }],
+      // Other rules...
+    },
+  };
+
